+++ app/helpers/MetaFormats/AnnotationConversion/AnnotationToAttributeConverter.php
@@ -1,0 +1,63 @@
+<?php
+
+namespace App\Helpers\MetaFormats\AnnotationConversion;
+
+use App\Helpers\MetaFormats\Attributes\Param;
+use App\Helpers\MetaFormats\Type;
+
+class AnnotationToAttributeConverter
+{
+    public static function convertFile(string $path)
+    {
+        $content = StandardAnnotationConverter::preprocessFile($path);
+        $nettePreprocess = NetteAnnotationConverter::regexReplaceAnnotations($content);
+
+        $netteCapturesList = $nettePreprocess["captures"];
+        $contentWithPlaceholders = $nettePreprocess["contentWithPlaceholders"];
+
+        // move the attribute lines below the comment block
+        $lines = [];
+        $netteAttributeLinesCount = 0;
+        $usingsAdded = false;
+        $paramAttributeClass = Utils::shortenClass(Param::class);
+        $paramTypeClass = Utils::shortenClass(Type::class);
+        foreach (Utils::fileStringToLines($contentWithPlaceholders) as $line) {
+            // detected the initial "use" block, add usings for new types
+            if (!$usingsAdded && strlen($line) > 3 && substr($line, 0, 3) === "use") {
+                $lines[] = "use App\\Helpers\\MetaFormats\\Attributes\\{$paramAttributeClass};";
+                $lines[] = "use App\\Helpers\\MetaFormats\\{$paramTypeClass};";
+                foreach (Utils::getValidatorNames() as $validator) {
+                    $lines[] = "use App\\Helpers\\MetaFormats\\Validators\\{$validator};";
+                }
+                // write the detected line (the first detected "use" line)
+                $lines[] = $line;
+                $usingsAdded = true;
+            // detected an attribute line placeholder, increment the counter and remove the line
+            } elseif (str_contains($line, NetteAnnotationConverter::$netteAttributePlaceholder)) {
+                $netteAttributeLinesCount++;
+            // detected the end of the comment block "*/", flush attribute lines
+            } elseif (trim($line) === "*/") {
+                $lines[] = $line;
+                for ($i = 0; $i < $netteAttributeLinesCount; $i++) {
+                    $annotationParameters = NetteAnnotationConverter::convertNetteRegexCapturesToDictionary($netteCapturesList[$i]);
+                    $parenthesesBuilder = NetteAnnotationConverter::convertRegexCapturesToParenthesesBuilder($annotationParameters);
+                    $attributeLine = "    #[{$paramAttributeClass}{$parenthesesBuilder->toString()}]";
+                    // change to multiline if the line is too long
+                    if (strlen($attributeLine) > 120) {
+                        $attributeLine = "    #[{$paramAttributeClass}{$parenthesesBuilder->toMultilineString(4)}]";
+                    }
+                    $lines[] = $attributeLine;
+                }
+
+                // remove the captures used in this endpoint
+                $netteCapturesList = array_slice($netteCapturesList, $netteAttributeLinesCount);
+                // reset the counters for the next detected endpoint
+                $netteAttributeLinesCount = 0;
+            } else {
+                $lines[] = $line;
+            }
+        }
+
+        return Utils::linesToFileString($lines);
+    }
+}
